Añadidos hashcode e equals en habitacion y residente y cambiado habitacion para que sea un objeto de tipo habitacion

+++ TuteladoJPA-1.1-src/TuteladoJPA-1.1/src/test/java/gei/id/tutelado/T03_Residentes.java
@@ -284,7 +284,7 @@ public class T03_Residentes {
         log.info("Probando gravacion de usuario con habitacion nulo ----------------------------------------------------");
         //produtorDatos.r1.setHabitacion(null);
         try {
-            produtorDatos.r1.setHabitacion((Integer) null);
+            produtorDatos.r1.setHabitacion(null);
             resDao.almacena(produtorDatos.r1);
             excepcion=false;
         } catch (Exception ex) {
